Apply usage of ServerContext to Server.kt

+++ src/main/kotlin/game/model/protobuf/equipment/PBEquipmentViewArmamentResponse.kt
@@ -8,7 +8,7 @@ import kotlinx.serialization.Serializable
  */
 @Serializable
 data class PBEquipmentViewArmamentResponse(
-    val playerId: Int,
+    val playerId: Long,
     val width: Int,
     val height: Int,
     val itemSlots: List<PBItemSlot>,
@@ -18,7 +18,7 @@ data class PBEquipmentViewArmamentResponse(
     val nextFreeUnlockLevel: Int?,
 ) {
     companion object {
-        fun dummy(pid: Int): PBEquipmentViewArmamentResponse {
+        fun dummy(pid: Long): PBEquipmentViewArmamentResponse {
             return PBEquipmentViewArmamentResponse(
                 playerId = pid,
                 width = 8,
